move more tests out of test_suite file
