+++ corehq/apps/app_manager/tests/__init__.py
@@ -28,6 +28,8 @@ try:
     from corehq.apps.app_manager.tests.test_case_meta import *
     from corehq.apps.app_manager.tests.test_case_list_form import *
     from corehq.apps.app_manager.tests.test_form_workflow import *
+    from corehq.apps.app_manager.tests.test_case_list_lookup import *
+    from corehq.apps.app_manager.tests.test_child_module import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
